457 dont disconnect on 503 - disconnects at start

+++ shared/src/commonMain/kotlin/uk/co/sentinelweb/cuer/app/ui/cast/CastController.kt
@@ -138,7 +138,7 @@ class CastController(
                     remoteNode?.name(),
                     isPlaying(),
                     if (isWatching()) {
-                        (volume / volumeMax * 100).toString() + " %"
+                        (volume / volumeMax * 100).toInt().toString() + " %"
                     } else {
                         "--"
                     },
